backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-community
langchain-groq
langchain-ollama
pydantic
python-dotenv
duckduckgo-search
beautifulsoup4
aiohttp
markdownify
sqlalchemy
aiosqlite
python-multipart
pyyaml
email-validator
httpx
